+++ src/main/java/com/apiit/eeashoppingapplication/Controllers/PaymentController.java
@@ -10,40 +10,52 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping(value = "/payment")
 public class PaymentController {
 
-        @Autowired
-        private PaymentRepository paymentRepository;
+    @Autowired
+    private PaymentRepository paymentRepository;
 
-        @GetMapping(path = "/all")
-        public @ResponseBody
-        Iterable<Payment> getPayments() { return paymentRepository.findAll();
-        }
+    @GetMapping(path = "/all")
+    public @ResponseBody Iterable<Payment> getPayments() {
+        return paymentRepository.findAll();
+    }
 
-        @GetMapping(path = "/{payment_id}")
-        public Payment getPayment(@PathVariable String payment_id) {
-            return paymentRepository.findById(payment_id).get();
-        }
+    @GetMapping(path = "/{payment_id}")
+    public Payment getPayment(@PathVariable String payment_id) {
+        return paymentRepository.findById(payment_id).get();
+    }
 
-        @PostMapping(path = "/new")
-        public Payment newPayment(@RequestBody Payment payment) {
-           System.out.println(payment.getPaymentName() + "done payment");
-           return paymentRepository.save(payment);
+    @PostMapping(path = "/new")
+    public Payment newPayment(@RequestBody Payment payment) {
 
-        }
+        Payment newpayment;
+
+        newpayment = paymentRepository.save(payment);
+        System.out.println(newpayment.getPaymentName() + "done payment");
+
+        return newpayment;
+    }
+
+
 
         @PutMapping
-        public Payment updatePromotions(@RequestBody Payment payment) {
+    public Payment updatePromotions(@RequestBody Payment payment) {
 
-            paymentRepository.save(payment);
+        paymentRepository.save(payment);
 
-            return payment;
-        }
-
-        @DeleteMapping(path = "/{payment_id}")
-        public void deletePayment(@PathVariable String payment_id) {
-
-            paymentRepository.deleteById(payment_id);
-
-        }
-
-
+        return payment;
     }
+
+    @DeleteMapping(path = "/{payment_id}")
+    public boolean deletePayment(@PathVariable String payment_id) {
+        boolean flag;
+        Payment payment = getPayment(payment_id);
+        if (payment != null) {
+            paymentRepository.deleteById(payment_id);
+            flag = true;
+        } else {
+            flag = false;
+
+        }
+        return flag;
+    }
+
+}
